+++ ui/apps/platform/cypress/integration/risk/podTimeline.test.js
@@ -147,10 +147,10 @@ describe('Risk Page Pod Event Timeline', () => {
             );
         });
 
-        it('should filter process activity events and whitelisted process activity events', () => {
+        it('should filter process activity events and process in baseline activity events', () => {
             openMockedPodEventTimelineView();
 
-            // the process activity + whitelisted process activity event should be visible
+            // the process activity + process in baseline activity event should be visible
             cy.get(selectors.eventTimeline.select.value).should('contain', 'Show All');
             cy.get(selectors.eventTimeline.timeline.mainView.event.processActivity);
             cy.get(selectors.eventTimeline.timeline.mainView.event.processInBaselineActivity);
@@ -161,7 +161,7 @@ describe('Risk Page Pod Event Timeline', () => {
                 `${selectors.eventTimeline.select.options}:eq(${FILTER_OPTIONS.POLICY_VIOLATIONS})`
             ).click({ force: true });
 
-            // the process activity + whitelisted process activity event should not be visible
+            // the process activity + process in baseline activity event should not be visible
             cy.get(selectors.eventTimeline.timeline.mainView.event.processActivity).should(
                 'not.exist'
             );
@@ -355,7 +355,7 @@ describe('Risk Page Pod Event Timeline', () => {
             );
         });
 
-        it('shows the whitelisted process activity event details', () => {
+        it('shows the process in baseline activity event details', () => {
             openMockedPodEventTimelineView();
 
             // trigger the tooltip
@@ -479,7 +479,7 @@ describe('Risk Page Pod Event Timeline', () => {
                 `${selectors.tooltip.legendContents}:eq(1) ${selectors.tooltip.legendContent.event.policyViolation}`
             );
 
-            // make sure the whitelisted process activity icon and text shows up
+            // make sure the process in baseline activity icon and text shows up
             cy.get(
                 `${selectors.tooltip.legendContents}:eq(2):contains("Baseline Process Activity")`
             );