+++ ui/apps/platform/src/Components/TimelineGraph/EventsGraph/ClusteredEventMarker/ClusteredEventMarker.test.js
@@ -105,7 +105,7 @@ test('should show a clustered process in baseline activity event marker', async 
             parentName: null,
             parentUid: -1,
             uid: 1000,
-            whitelisted: true,
+            inBaseline: true,
             timestamp: '2020-04-20T20:20:20.358227916Z',
         };
     });